CamelCased TableSchema module
