+++ .travis.yml
@@ -2,7 +2,7 @@
 language: julia
 os:
   - linux
-  - osx
+  # - osx
 julia:
   - 0.6
   - nightly
@@ -31,5 +31,3 @@ git:
 after_success:
   # push coverage results to Coveralls
   - julia -e 'cd(Pkg.dir("TableSchema")); Pkg.add("Coverage"); using Coverage; Coveralls.submit(Coveralls.process_folder())'
-  # push coverage results to Codecov
-  - julia -e 'cd(Pkg.dir("TableSchema")); Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())'
