+++ README.md
@@ -1,1 +1,5 @@
 # TableSchema
+
+A Julia library for working with [Table Schema](http://specs.frictionlessdata.io/).
+
+:construction: Work In Progress.
