+++ test/runtests.jl
@@ -1,5 +1,32 @@
 using TableSchema
 using Base.Test
 
-# write your own tests here
-@test 1 == 2
+include("data.jl")
+
+@testset "Loading a minimal schema" begin
+    s = TableSchema.load(DESCRIPTOR_MIN)
+    @test length(s.fields) == 2
+    d1 = s.fields[1].descriptor
+    @test d1._name == "id"
+    @test d1._type == "string"
+    d2 = s.fields[2].descriptor
+    @test d2._type == "integer"
+    @test d2._required == false
+end
+
+@testset "Loading a schema from JSON" begin
+    s = TableSchema.load(DESCRIPTOR_MIN_JSON)
+    @test length(s.fields) == 2
+    d1 = s.fields[1].descriptor
+    @test d1._name == "id"
+    d2 = s.fields[2].descriptor
+    @test d2._required == false
+end
+
+@testset "Loading a complex schema" begin
+    s = TableSchema.load(DESCRIPTOR_MAX_JSON)
+    @test length(s.fields) == 5
+    @test length(s.primaryKey) == 1
+    @test length(s.foreignKeys) == 1
+    @test length(s.missingValues) == 3
+end
